Fix picture failures in dark lighting
After sending the GET_PICTURE command the code waits for a response
indicating that there is picture data ready to be read. The delay
between this command and the response is typically 800ms, but when
taking a picture in the dark this delay can be larger than 1200ms.
This patch increases the read timeout from 1s to 5s to ensure that
pictures are taken reliably, even in dark conditions.